a 	modified:   ContarBits.cpp

+++ Exercicios/ContarBits.cpp
@@ -1,9 +1,9 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main(){
-    int n;
+    long long int n;
     cin >> n;
-    int cont = 1; 
+    long long int cont = 1; 
     
     while(1){
     	if(n % 2 == 1) cont++;
